Move perf to utils
